@Wayoung7 2024-04-30
Yeah, deque is faster and more suitable for this use case. I didn't even think of `VecDeque` at that time when I was programing this. Thanks for your suggestion.